Add ErrorFactory, add documentation, fix style
