+++ src/Tizen.WsShell.Quickpanel/Tizen.WsShell.Quickpanel/Interop/Interop.Libraries.cs
@@ -1,10 +1,26 @@
-﻿internal static partial class Interop
+﻿//Copyright 2018 Samsung Electronics Co., Ltd
+//
+//Licensed under the Apache License, Version 2.0 (the "License");
+//you may not use this file except in compliance with the License.
+//You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+//Unless required by applicable law or agreed to in writing, software
+//distributed under the License is distributed on an "AS IS" BASIS,
+//WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//See the License for the specific language governing permissions and
+//limitations under the License.
+
+internal static partial class Interop
 {
     internal static partial class Libraries
     {
         public const string Quickpanel = "libtzsh_quickpanel.so.0";
 
         public const string TzshCommon = "libtzsh_common.so.0";
+
+        public const string Elementary = "libelementary.so.1";
     }
 }
