+++ src/Tizen.WsShell.Quickpanel/Tizen.WsShell.Quickpanel/Interop/Interop.WsShell.cs
@@ -1,8 +1,19 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Runtime.InteropServices;
-using System.Text;
+﻿//Copyright 2018 Samsung Electronics Co., Ltd
+//
+//Licensed under the Apache License, Version 2.0 (the "License");
+//you may not use this file except in compliance with the License.
+//You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+//Unless required by applicable law or agreed to in writing, software
+//distributed under the License is distributed on an "AS IS" BASIS,
+//WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//See the License for the specific language governing permissions and
+//limitations under the License.
 
+using System;
+using System.Runtime.InteropServices;
 
 internal static partial class Interop
 {
@@ -14,8 +25,7 @@ internal static partial class Interop
         [DllImport(Libraries.Quickpanel, EntryPoint = "tzsh_quickpanel_create")]
         internal static extern int Destroy(IntPtr tzsh);
 
-        //TODO: FIND BETTER PLACE FOR THIS
-        [DllImport("libelementary.so.1", EntryPoint = "elm_win_window_id_get")]
+        [DllImport(Libraries.Elementary, EntryPoint = "elm_win_window_id_get")]
         internal static extern int GetElmWindowId(IntPtr win);
     }
 }
